break-up builders.go into many files because it was too big
